pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp10x"
version = "0.1.0"
description = "MCP server for 10x engineering — Jira, Confluence, coding rules, decision log, and workflow prompts"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0.0",
    "atlassian-python-api>=3.41.0",
    "pyyaml>=6.0",
    "markdownify>=0.13.0",
    "markdown>=3.5",
    "python-gitlab>=4.0.0",
]

[project.scripts]
mcp10x-server = "mcp10x.server:main"

[tool.setuptools.packages.find]
where = ["src"]
